requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
decorator==5.1.1
distro==1.9.0
forex-python==1.8
future==1.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huepy==1.2.1
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
iniconfig==2.0.0
instabot==0.117.0
jiter==0.8.2
mock==5.1.0
moviepy==2.1.1
numpy==2.2.1
openai==1.58.1
packaging==24.2
pillow==10.4.0
pluggy==1.5.0
proglog==0.1.10
pydantic==2.10.4
pydantic_core==2.27.2
pydub==0.25.1
PySocks==1.7.1
pytest==8.3.4
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
responses==0.25.3
schedule==1.2.2
setuptools==75.6.0
simplejson==3.19.3
six==1.17.0
sniffio==1.3.1
tiktoken==0.8.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
